Named parameters, generics, separate classes
Prasmingai panaudojau named parameters su expense įrašymu.
Panaudojau generics su expense listu.
Vietoj vienos DBmanager klasės sukūriau atskiras klases expense, income ir goal
